(cleanup) Finish moving files into appropriate folders
