+++ client/components/TopNavBar/Friends/FriendItemList.js
